+++ contracts/multisig-prover/src/execute.rs
@@ -13,12 +13,12 @@ use router_api::{ChainName, CrossChainId, Message};
 use service_registry::state::WeightedWorker;
 
 use crate::{
-    command::{Command, Payload},
     contract::START_MULTISIG_REPLY_ID,
     encoding::{make_operators, CommandBatchBuilder},
     error::ContractError,
+    payload::Payload,
     state::{
-        Config, COMMAND, COMMANDS_BATCH, CONFIG, CURRENT_WORKER_SET, NEXT_WORKER_SET, REPLY_BATCH,
+        Config, COMMANDS_BATCH, CONFIG, CURRENT_WORKER_SET, NEXT_WORKER_SET, PAYLOAD, REPLY_BATCH,
     },
     types::{BatchId, WorkersInfo},
 };
@@ -231,12 +231,13 @@ pub fn update_worker_set(deps: DepsMut, env: Env) -> Result<Response, ContractEr
 
             save_next_worker_set(deps.storage, &new_worker_set)?;
 
-            let command = Command::new(Payload::WorkerSet(new_worker_set));
-            COMMAND.save(deps.storage, &command.id, &command)?;
-            REPLY_BATCH.save(deps.storage, &command.id)?;
+            let payload = Payload::WorkerSet(new_worker_set);
+            let payload_id = payload.id();
+            PAYLOAD.save(deps.storage, &payload_id, &payload)?;
+            REPLY_BATCH.save(deps.storage, &payload_id)?;
 
             let msg_digest =
-                command.digest(config.encoder, &config.domain_separator, &cur_worker_set);
+                payload.digest(config.encoder, &config.domain_separator, &cur_worker_set);
 
             let start_sig_msg = multisig::msg::ExecuteMsg::StartSigningSession {
                 worker_set_id: cur_worker_set.id(),
