+++ contracts/multisig-prover/src/lib.rs
@@ -1,10 +1,10 @@
-pub mod command;
 pub mod contract;
 pub mod encoding;
 pub mod error;
 pub mod events;
 mod execute;
 pub mod msg;
+pub mod payload;
 mod query;
 mod reply;
 pub mod state;
