+++ contracts/multisig-prover/src/reply.rs
@@ -1,7 +1,7 @@
 use cosmwasm_std::{from_binary, DepsMut, Reply, Response, Uint64};
 use cw_utils::{parse_reply_execute_data, MsgExecuteContractResponse};
 
-use crate::state::{COMMAND, CONFIG};
+use crate::state::{CONFIG, PAYLOAD};
 use crate::{
     error::ContractError,
     events::Event,
@@ -26,7 +26,7 @@ pub fn start_multisig_reply(deps: DepsMut, reply: Reply) -> Result<Response, Con
                 &command_batch_id,
             )?;
 
-            let msg_ids = COMMAND.load(deps.storage, &command_batch_id)?.message_ids();
+            let msg_ids = PAYLOAD.load(deps.storage, &command_batch_id)?.message_ids();
 
             Ok(Response::new().add_event(
                 Event::ProofUnderConstruction {
