+++ contracts/multisig-prover/src/types.rs
@@ -28,7 +28,7 @@ impl Display for CommandType {
     }
 }
 
-#[deprecated(since = "0.2.0", note = "Please use command::Payload instead")]
+#[deprecated(since = "0.2.0", note = "Please use payload module instead")]
 #[cw_serde]
 pub struct Command {
     pub id: HexBinary,
@@ -85,7 +85,7 @@ impl BatchId {
     }
 }
 
-#[deprecated(since = "0.2.0", note = "Please use command::Command instead")]
+#[deprecated(since = "0.2.0", note = "Please use payload module instead")]
 #[cw_serde]
 pub struct CommandBatch {
     pub id: BatchId,
